requirements.txt
appdirs==1.4.3
asgiref==3.2.3
attrs==19.3.0
black==19.10b0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.4
django-cors-headers==3.2.1
django-filter==2.2.0
django-oauth-toolkit==1.3.0
djangorestframework==3.11.0
djangorestframework-gis==0.15
djangorestframework-jsonapi==3.1.0
drf-extensions==0.6.0
drf-yasg==1.17.1
entrypoints==0.3
flake8==3.7.9
idna==2.8
inflection==0.3.1
itypes==1.1.0
Jinja2==2.11.1
Markdown==3.2.1
MarkupSafe==1.1.1
mccabe==0.6.1
oauthlib==3.1.0
packaging==20.3
pathspec==0.6.0
psycopg2-binary==2.8.4
pycodestyle==2.5.0
pyflakes==2.1.1
pyparsing==2.4.6
pytz==2019.3
PyYAML==5.3.1
regex==2019.12.20
requests==2.22.0
rope==0.16.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.0
toml==0.10.0
typed-ast==1.4.0
uritemplate==3.0.1
urllib3==1.25.8
uWSGI==2.0.18
